src/backend/app/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "parvaz_academy backend DjangoRestFramework"
authors = ["AlirezMortezaei <Mortezaei2324@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
django-dotenv = "^1.4.2"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
drf-spectacular = "^0.27.1"
pillow = "^10.2.0"
redis = "^5.0.1"
kavenegar = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
